pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bor-risk-agent"
version = "0.1.0"
description = "Supply-chain risk analysis agent using LangGraph"
requires-python = ">=3.10"
dependencies = [
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "langchain-openai>=0.3",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.hatch.build.targets.wheel]
packages = ["src/bor_risk"]

[tool.pytest.ini_options]
testpaths = ["tests"]
